rag-yandex-bot/requirements.txt
# Yandex Bot Service Dependencies

# HTTP client
aiohttp==3.11.0

# Environment variables
python-dotenv==1.0.1

qdrant-client==1.9.0
openai
